add s2j call for dlopen
