+++ src/tool/s2jgo/common.go
@@ -1,0 +1,62 @@
+package s2jgo
+
+import "unsafe"
+
+type Result struct {
+	Stdout []byte
+	Stderr []byte
+}
+
+type argHolder struct {
+	argvVec []uintptr
+	cStrs   [][]byte
+}
+
+type outData struct {
+	cb            func(data []byte, isStdErr bool)
+	stderrCapture []byte
+}
+
+func callback(data unsafe.Pointer, size uintptr, isStdErr uintptr, ctx unsafe.Pointer) uintptr {
+	out := (*outData)(ctx)
+	if out.cb != nil {
+		out.cb(unsafe.Slice((*byte)(data), size), isStdErr != 0)
+	}
+	if isStdErr != 0 {
+		out.stderrCapture = append(out.stderrCapture, unsafe.Slice((*byte)(data), size)...)
+	}
+	return 0
+}
+
+func (a *argHolder) makeArg(args []string) (argc uintptr, argv uintptr) {
+	a.argvVec = make([]uintptr, len(args)+1)
+	a.cStrs = make([][]byte, len(args))
+	for i, arg := range args {
+		a.cStrs[i] = append([]byte(arg), 0)
+	}
+	for i := range a.cStrs {
+		a.argvVec[i] = uintptr(unsafe.Pointer(unsafe.SliceData(a.cStrs[i])))
+	}
+	a.argvVec[len(args)] = 0
+	argc = uintptr(len(args))
+	argv = uintptr(unsafe.Pointer(unsafe.SliceData(a.argvVec)))
+	return
+}
+
+func (s *Src2JSON) Call(args []string, cap Capability) (*Result, error) {
+	var stdout, stderr []byte
+	err := s.CallIOCallback(args, cap, func(data []byte, isStdErr bool) {
+		if isStdErr {
+			stderr = append(stderr, data...)
+		} else {
+			stdout = append(stdout, data...)
+		}
+	})
+	if err != nil {
+		return nil, err
+	}
+	return &Result{
+		Stdout: stdout,
+		Stderr: stderr,
+	}, nil
+}
